+++ module-api/src/main/java/com/kernel360/ModuleApiApplication.java
@@ -1,0 +1,13 @@
+package com.kernel360;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModuleApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ModuleApiApplication.class, args);
+    }
+
+}
