+++ module-api/src/main/java/com/kernel360/member/controller/MemberController.java
@@ -1,0 +1,5 @@
+package com.kernel360.member.controller;
+
+
+public class MemberController {
+}
