+++ module-api/src/main/java/com/kernel360/member/dto/MemberRequest.java
@@ -1,0 +1,4 @@
+package com.kernel360.member.dto;
+
+public class MemberRequest {
+}
