+++ module-api/src/main/java/com/kernel360/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.kernel360.member.service;
+
+public class MemberService {
+}
