+++ module-api/src/test/java/com/kernel360/ModuleApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.kernel360;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ModuleApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
