+++ module-api/src/test/resources/application.yml
@@ -1,0 +1,10 @@
+spring:
+  datasource:
+    driver-class-name: org.postgresql.Driver
+    url: jdbc:postgresql://localhost:5431/washpedia
+    username: postgres
+    password:
+
+jasypt:
+  encryptor:
+    algorithm: PBEWithMD5AndTripleDES
