+++ module-domain/build.gradle
@@ -11,13 +11,30 @@ java {
     sourceCompatibility = '17'
 }
 
+configurations {
+    compileOnly {
+        extendsFrom annotationProcessor
+    }
+}
+
 repositories {
     mavenCentral()
 }
 
 dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter'
 }
 
 tasks.named('test') {
     useJUnitPlatform()
 }
+
+tasks.bootJar {
+    enabled = false
+}
+
+tasks.jar {
+    enabled = true
+}
+
+tasks.register("prepareKotlinBuildScriptModel") {}
