+++ module-domain/src/main/java/com/kernel360/base/BaseEntity.java
@@ -1,0 +1,4 @@
+package com.kernel360.base;
+
+public class BaseEntity {
+}
