+++ module-domain/src/main/java/com/kernel360/member/entity/Member.java
@@ -1,0 +1,4 @@
+package com.kernel360.member.entity;
+
+public class Member {
+}
