+++ module-domain/src/main/java/com/kernel360/member/repository/MemberRepository.java
@@ -1,0 +1,4 @@
+package com.kernel360.member.repository;
+
+public class MemberRepository {
+}
