+++ settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'washpedia'
+
+include 'module-api'
+include 'module-common'
+include 'module-domain'
