--- src/main/java/com/kernel360/washpedia/WashpediaApplication.java
@@ -1,13 +1,0 @@
-package com.kernel360.washpedia;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WashpediaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(WashpediaApplication.class, args);
-	}
-
-}
