--- src/main/java/com/kernel360/washpedia/domain/member/controller/MemberController.java
@@ -1,5 +1,0 @@
-package com.kernel360.washpedia.domain.member.controller;
-
-
-public class MemberController {
-}
