--- src/main/java/com/kernel360/washpedia/domain/member/dto/MemberRequest.java
@@ -1,4 +1,0 @@
-package com.kernel360.washpedia.domain.member.dto;
-
-public class MemberRequest {
-}
