--- src/main/java/com/kernel360/washpedia/domain/member/entity/Member.java
@@ -1,4 +1,0 @@
-package com.kernel360.washpedia.domain.member.entity;
-
-public class Member {
-}
