--- src/main/java/com/kernel360/washpedia/domain/member/repository/MemberRepository.java
@@ -1,4 +1,0 @@
-package com.kernel360.washpedia.domain.member.repository;
-
-public class MemberRepository {
-}
