--- src/main/java/com/kernel360/washpedia/domain/member/service/MemberService.java
@@ -1,4 +1,0 @@
-package com.kernel360.washpedia.domain.member.service;
-
-public class MemberService {
-}
