--- src/main/java/com/kernel360/washpedia/global/domain/BaseEntity.java
@@ -1,4 +1,0 @@
-package com.kernel360.washpedia.global.domain;
-
-public class BaseEntity{
-}
