--- src/test/java/com/kernel360/washpedia/WashpediaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kernel360.washpedia;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WashpediaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
