--- src/test/resources/application.yml
@@ -1,12 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://localhost:5431/washpedia
-    username: postgres
-    password: 1q2w3e4r
-
-jasypt:
-  encryptor:
-    algorithm: PBEWithMD5AndTripleDES
-
-
